514787: A bit of sensible renaming
