+++ src/config/nunjucks/index.js
@@ -5,9 +5,9 @@ import nunjucks from 'nunjucks'
 import Vision from '@hapi/vision'
 
 import { config } from '~/src/config/config.js'
-import { context } from '~/src/config/nunjucks/context/index.js'
-import * as filters from '~/src/config/nunjucks/filters/index.js'
-import * as globals from '~/src/config/nunjucks/globals/index.js'
+import { context } from '~/src/config/nunjucks/context/context.js'
+import * as filters from '~/src/config/nunjucks/filters/filters.js'
+import * as globals from '~/src/config/nunjucks/globals/globals.js'
 
 const dirname = path.dirname(fileURLToPath(import.meta.url))
