+++ src/server/admin/permissions/controllers/edit/permission-details-form.js
@@ -1,7 +1,7 @@
 import Boom from '@hapi/boom'
 
 import Joi from '~/src/server/common/helpers/extended-joi.js'
-import { fetchScope } from '~/src/server/admin/permissions/helpers/fetchers.js'
+import { fetchPermissionsScope } from '~/src/server/admin/permissions/helpers/fetchers.js'
 import { buildOptions } from '~/src/server/common/helpers/options/build-options.js'
 
 const editPermissionDetailsFormController = {
@@ -14,7 +14,10 @@ const editPermissionDetailsFormController = {
     }
   },
   handler: async (request, h) => {
-    const { scope } = await fetchScope(request, request.params.scopeId)
+    const { scope } = await fetchPermissionsScope(
+      request,
+      request.params.scopeId
+    )
     const kindOptions = buildOptions(
       [
         { text: 'User', value: 'user' },
