+++ src/server/admin/permissions/controllers/permission.js
@@ -1,8 +1,8 @@
 import Boom from '@hapi/boom'
 
 import Joi from '~/src/server/common/helpers/extended-joi.js'
-import { formatText } from '~/src/config/nunjucks/filters/index.js'
-import { fetchScope } from '~/src/server/admin/permissions/helpers/fetchers.js'
+import { formatText } from '~/src/config/nunjucks/filters/filters.js'
+import { fetchPermissionsScope } from '~/src/server/admin/permissions/helpers/fetchers.js'
 import { transformScopeToSummary } from '~/src/server/admin/permissions/transformers/scope-to-summary.js'
 import { transformScopeTeamsToTaskList } from '~/src/server/admin/permissions/transformers/scope-teams-to-task-list.js'
 import { transformScopeUsersToTaskList } from '~/src/server/admin/permissions/transformers/scope-users-to-task-list.js'
@@ -17,7 +17,10 @@ const permissionController = {
     }
   },
   handler: async (request, h) => {
-    const { scope } = await fetchScope(request, request.params.scopeId)
+    const { scope } = await fetchPermissionsScope(
+      request,
+      request.params.scopeId
+    )
     const formattedValue = formatText(scope.value)
 
     return h.view('admin/permissions/views/permission', {
