+++ src/server/running-services/controllers/running-services-list.js
@@ -3,7 +3,7 @@ import Boom from '@hapi/boom'
 
 import { provideAuthedUser } from '~/src/server/common/helpers/auth/pre/provide-authed-user.js'
 import { buildRunningServicesTableData } from '~/src/server/running-services/helpers/build-running-services-table-data.js'
-import { formatText } from '~/src/config/nunjucks/filters/index.js'
+import { formatText } from '~/src/config/nunjucks/filters/filters.js'
 
 const runningServicesListController = {
   options: {
