+++ src/server/services/proxy/controllers/environment.js
@@ -1,6 +1,6 @@
 import Boom from '@hapi/boom'
 
-import { formatText } from '~/src/config/nunjucks/filters/index.js'
+import { formatText } from '~/src/config/nunjucks/filters/filters.js'
 import { provideService } from '~/src/server/services/helpers/pre/provide-service.js'
 import { serviceParamsValidation } from '~/src/server/services/helpers/schema/service-params-validation.js'
 import { findProxyRulesForEnvironment } from '~/src/server/services/proxy/helpers/find-proxy-rules.js'
