+++ src/server/services/secrets/controllers/environment.js
@@ -1,6 +1,6 @@
 import Boom from '@hapi/boom'
 
-import { formatText } from '~/src/config/nunjucks/filters/index.js'
+import { formatText } from '~/src/config/nunjucks/filters/filters.js'
 import { fetchSecrets } from '~/src/server/common/helpers/fetch/fetch-secrets.js'
 import { provideService } from '~/src/server/services/helpers/pre/provide-service.js'
 import { environmentSecrets } from '~/src/server/services/secrets/transformers/environment-secrets.js'
